created PlayerCard component to make switch statement to update user info
