+++ client/src/pages/GameHistory.jsx
@@ -4,10 +4,12 @@ import { useState, useEffect } from "react"
 import { useQuery, useMutation } from "@apollo/client"
 import PlayerForm from "../components/PlayerForm"
 import TokenList from "../assets/tokenList.json"
-import PropertiesCard from "../components/PropertiesCard"
+// import PropertiesCard from "../components/PropertiesCard"
+import PlayerCard from "../components/PlayerCard"
 
 const GameHistory = () => {
     // TODO:
+    // MAKE SWITCH STATEMENT TO SHOW AN UPDATE FORM AND STATIC MESSAGE FOR EACH USER
     // TEST OUT UPDATE QUERY
     // MAKE SURE QUERY WORK WITH PROPERTIE
     // NOTE:
@@ -20,7 +22,7 @@ const GameHistory = () => {
         fetchPolicy: "no-cache"
     })
     const [deleteGame,{error}] = useMutation(DELETE_GAME)
-    const [removeOnePlayerFromGame, {erro}] = useMutation(REMOVE_ONE_PLAYER_FROM_GAME)
+    // const [removeOnePlayerFromGame, {erro}] = useMutation(REMOVE_ONE_PLAYER_FROM_GAME)
     const [upddatePlayerInfo, {err}] = useMutation(UPDATE_PLAYER_INFO)
     const gameList = data?.me?.gameMaster || [];
     const currentProperties = data2?.data?.allProperties || []
@@ -32,13 +34,13 @@ const GameHistory = () => {
     let [playersLength, setPlayersLenth] = useState(0)
     let [avialableTokens, setAvialableTokens] = useState([])
 
-    const updatePlayerFunc = (event)=>{
-        event.preventDefault()
-        console.log(event.target.parentNode.querySelector('.playerMoney').value)
-        // CREATE A VARIABLE TO STORE UPDATED MONEY 
+    // const updatePlayerFunc = (event)=>{
+    //     event.preventDefault()
+    //     console.log(event.target.parentNode.querySelector('.playerMoney').value)
+    //     // CREATE A VARIABLE TO STORE UPDATED MONEY 
 
-        // CREATE A VARIABLE TO STORE UPDATED POSITION
-    }
+    //     // CREATE A VARIABLE TO STORE UPDATED POSITION
+    // }
 
     const onAddBtnClick = event => {
         playersLength = playersLength + 1
@@ -90,18 +92,18 @@ const GameHistory = () => {
         }
 
     }
-    const deletePlayer = async (event)=>{
-        let playerId = event.target.getAttribute('data-playerid')
-        let gameId = currentGameId
-        try{
-            const {data} = await removeOnePlayerFromGame({
-                variables:{gameId,playerId}
-            })
-            window.location.reload()
-        }catch(err){
-            console.error(err)
-        }
-    }
+    // const deletePlayer = async (event)=>{
+    //     let playerId = event.target.getAttribute('data-playerid')
+    //     let gameId = currentGameId
+    //     try{
+    //         const {data} = await removeOnePlayerFromGame({
+    //             variables:{gameId,playerId}
+    //         })
+    //         window.location.reload()
+    //     }catch(err){
+    //         console.error(err)
+    //     }
+    // }
     const getAvailableProperties =  (game, currentProperties )=>{
         if(game && currentProperties){
         let playerProperties = game.map((player)=>{
@@ -123,15 +125,12 @@ const GameHistory = () => {
     }
     function getAvaliableTokens(currenPlayer, allTokens){
         if(currenPlayer && allTokens){
-            console.log(currenPlayer)
-            console.log(allTokens)
             // LOOK INTO THIS LATER
             let filteredToken = allTokens.filter(function(jsonToken){
                 return !currenPlayer.find(function(playerToken){
                     return jsonToken.tokenName == playerToken.token
                 })
             });
-            console.log(filteredToken)
             setAvialableTokens(filteredToken)
         }
     }
@@ -176,7 +175,8 @@ const GameHistory = () => {
                     {selectedGame[0].savedPlayers?.map((player)=>{
                         return (
                             <li key={player._id}>
-                                <form>
+                                <PlayerCard player={player} currentGameId={currentGameId} takenProperties={takenProperties}/>
+                                {/* <form>
                                     <p>Name: {player.name}</p>
                                     <p>Token: {player.token}</p>
                                     <label>Money</label>
@@ -198,7 +198,7 @@ const GameHistory = () => {
                                     <br></br>
                                     <button data-playerid={player._id} onClick={(event) => {updatePlayerFunc(event)}}>Update</button>
                                     <button data-playerid={player._id} onClick={(event)=> deletePlayer(event)}>Delete</button>
-                                </form>
+                                </form> */}
                             </li>
                         )
                     })}
